+++ burst-kotlin-plugin-tests/src/test/java/app/cash/burst/kotlin/IrDumpTestGenerated.java
@@ -35,5 +35,11 @@ public class IrDumpTestGenerated extends AbstractIrDumpTest {
     public void testFunctionParameters() {
       runTest("burst-kotlin-plugin-tests/src/test/data/dump/ir/FunctionParameters.kt");
     }
+
+    @Test
+    @TestMetadata("TestInterceptor.kt")
+    public void testTestInterceptor() {
+      runTest("burst-kotlin-plugin-tests/src/test/data/dump/ir/TestInterceptor.kt");
+    }
   }
 }